use GetWorkdirFromManifestPath from filesystem
