+++ cmd/stack/deploy.go
@@ -15,6 +15,7 @@ package stack
 
 import (
 	"context"
+	"github.com/okteto/okteto/pkg/filesystem"
 	"os"
 	"runtime"
 	"strings"
@@ -62,7 +63,7 @@ func deploy(ctx context.Context, at, insights buildTrackerInterface, ioCtrl *io.
 
 			options.StackPaths = loadComposePaths(options.StackPaths)
 			if len(options.StackPaths) == 1 {
-				workdir := model.GetWorkdirFromManifestPath(options.StackPaths[0])
+				workdir := filesystem.GetWorkdirFromManifestPath(options.StackPaths[0])
 				if err := os.Chdir(workdir); err != nil {
 					return err
 				}
@@ -93,7 +94,7 @@ func deploy(ctx context.Context, at, insights buildTrackerInterface, ioCtrl *io.
 	cmd.Flags().BoolVarP(&options.ForceBuild, "build", "", false, "build images before starting any compose service")
 	cmd.Flags().BoolVarP(&options.Wait, "wait", "", false, "wait until a minimum number of containers are in a ready state for every service")
 	cmd.Flags().BoolVarP(&options.NoCache, "no-cache", "", false, "do not use cache when building the image")
-	cmd.Flags().DurationVarP(&options.Timeout, "timeout", "t", (10 * time.Minute), "the length of time to wait for completion, zero means never. Any other values should contain a corresponding time unit e.g. 1s, 2m, 3h ")
+	cmd.Flags().DurationVarP(&options.Timeout, "timeout", "t", 10*time.Minute, "the length of time to wait for completion, zero means never. Any other values should contain a corresponding time unit e.g. 1s, 2m, 3h ")
 	cmd.Flags().StringVarP(&options.Progress, "progress", "", oktetoLog.TTYFormat, "show plain/tty build output (default \"tty\")")
 	return cmd
 }
